Parse event input data based on column headers #6
Just needed to avoid replacing of header data to make columns sparse
as it appears the column headers are in the default format csv files.
This will allow a handmade csv file with just the needed columns
(which are:  FirstName, LastName, Email, CellPhone, ROLE and Room)
Also, fixed catching some parsing errors on misformatted data files
which could crash the backend!