Add event data download support #5
Added ability to archive the event to a JSON file.  This is at least
easy to read in a browser - if a csv or other format file is desired,
I'll make further changes down the road.  Also, removed write-json-file
package that I had accidentally included in the last package update.